Refactor: Generate Images and Text
