+++ img2xl_app/urls.py
@@ -19,11 +19,13 @@ urlpatterns = [
     url(r'^export/(?P<result_id>\d+)/$', views.export, name='export'),
     url(r'^export_to_sheets/(?P<result_id>\d+)/$', views.export_to_sheets, name='export_to_sheets'),
     url(r'^result/(?P<result_id>\d+)/update/$', views.update_table_data, name='update_table_data'),
+    
+    # Trích xuất text với AI
     url(r'^api/generate-ai-content/$', views.generate_ai_content, name='generate_ai_content'),
 
     # Trích xuất ảnh với gemini
-    url(r'^extract-only-api/$', views.extract_only_api, name='extract_only_api'),
-    url(r'^batch-extract-api/$', views.batch_extract_api, name='batch_extract_api'),
+    url(r'^extract-only-api/$', views.generate_ai_images, name='extract_only_api'),
+    url(r'^batch-extract-api/$', views.generate_ai_images, name='batch_extract_api'),
 
     url(r'^tasks/auto-cleanup/$', views.auto_cleanup_task, name='cleanup_old_data'),
     url(r'^documents/$', views.documents_view, name='documents'),
